app/src/androidTest/java/com/example/note/presentation/ui/notes/NotesScreenTest.kt
package com.example.note.presentation.ui.notes

import androidx.compose.ui.test.assertCountEquals
import androidx.compose.ui.test.assertIsDisplayed
import androidx.compose.ui.test.junit4.createAndroidComposeRule
import androidx.compose.ui.test.onAllNodesWithContentDescription
import androidx.compose.ui.test.onNodeWithContentDescription
import androidx.hilt.navigation.compose.hiltViewModel
import com.example.note.mock.MockSetup
import com.example.note.presentation.ui.MainActivity
import dagger.hilt.android.testing.HiltAndroidRule
import dagger.hilt.android.testing.HiltAndroidTest
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import javax.inject.Inject

@HiltAndroidTest
class NotesScreenTest {

    @get:Rule(order = 0)
    val hiltRule = HiltAndroidRule(this)

    @get:Rule(order = 1)
    val composeRule = createAndroidComposeRule<MainActivity>()

    @Inject
    lateinit var mockSetup: MockSetup

    @Before
    fun setUp() {
        hiltRule.inject()
        mockSetup.init()
        composeRule.setContent {
            val notesViewModel = hiltViewModel<NotesViewModel>()
            NotesScreen(
                viewModel = notesViewModel,
                navigateToDetail = {},
                navigateToSettings = {}
            )
        }
    }

    @Test
    fun assertIcons_isVisible() {
        composeRule.onNodeWithContentDescription("Search view").assertIsDisplayed()
        composeRule.onNodeWithContentDescription("Settings").assertIsDisplayed()
        composeRule.onNodeWithContentDescription("Add note").assertIsDisplayed()
        composeRule.onAllNodesWithContentDescription("Note card").assertCountEquals(6)
    }
}